Updates required after celery executor move to providers
