+++ airflow/stats.py
@@ -22,7 +22,6 @@ import socket
 from typing import TYPE_CHECKING, Callable
 
 from airflow.configuration import conf
-from airflow.metrics import datadog_logger, otel_logger, statsd_logger
 from airflow.metrics.base_stats_logger import NoStatsLogger, StatsLogger
 
 log = logging.getLogger(__name__)
@@ -46,10 +45,16 @@ class _Stats(type):
         if not hasattr(cls.__class__, "factory"):
             is_datadog_enabled_defined = conf.has_option("metrics", "statsd_datadog_enabled")
             if is_datadog_enabled_defined and conf.getboolean("metrics", "statsd_datadog_enabled"):
+                from airflow.metrics import datadog_logger
+
                 cls.__class__.factory = datadog_logger.get_dogstatsd_logger
             elif conf.getboolean("metrics", "statsd_on"):
+                from airflow.metrics import statsd_logger
+
                 cls.__class__.factory = statsd_logger.get_statsd_logger
             elif conf.getboolean("metrics", "otel_on"):
+                from airflow.metrics import otel_logger
+
                 cls.__class__.factory = otel_logger.get_otel_logger
             else:
                 cls.__class__.factory = NoStatsLogger
